SHA256Helper

+++ src/Zek/Contracts/FileDto.cs
@@ -4,12 +4,13 @@
     {
         public TId Id { get; set; }
         public string FileName { get; set; }
-        public int? FileSize { get; set; }
+        public long? FileSize { get; set; }
         public string Extension { get; set; }
         public string MediaType { get; set; }
         public TContent Content { get; set; }
         public bool? IsDeleted { get; set; }
     }
+
     public class FileDto<TId> : FileDto<TId, byte[]>
     {
     }
